requirements.txt
asgiref==3.8.1
Django==5.1.4
mysql-connector-python==9.1.0
mysqlclient==2.2.6
sqlparse==0.5.3
